Solution 1
DX11Base is now the parent class of RenderingManager rather than Game.
Reasoning:
-In the book, making DX11Base a parent of "CubeDemo()" worked because it
was also the class that handled rendering. Due to having a rendering
manager this was the easy way to get past this.
-Having it be the parent class of RenderingManager still means that it
will be separated from the rest of the code
-While it's still separated from the rest of the code, I can easily
reference it during rendering so there will be no issues.
-100% will not cause Circular Dependency issues.